API: prevent "zero" timestamp on ticket creation
In Go, an un-initialized `time.Time` has a value of "0" which the
GraphQL server will accept (as it is not `nil`) and PostgreSQL will
write it to the database. However, the API will afterwards fail to
return the creation date for this ticket, as the value is considered
null, but the ticket type defined to not allow null.

As a first step, prevent clients from creating tickets with the "0"
timestamp by adding a validation that checks the value.

References: https://todo.sr.ht/~sircmpwn/todo.sr.ht/259