todosrht: Fix 405 error when clicking 'next' in ticket list
When using the ticket preview feature, the URL would previously change
to /~user/tracker/submit. Attempting to navigate to the next page of
tickets afterwards would cause a 405 error since
/~user/tracker/submit?page=N is an invalid link.

Fix by changing the ticket submit route to /~user/tracker, so that
adding ?page=N results in a valid link.