api: send event notifications via meta.sr.ht GQL
This commit implements the sending of event notification emails via the
meta.sr.ht GraphQL API, which will cause the emails to be properly
signed and potentially encrypted (according to the recipients privacy
settings).

Almost the entire SendEmail function had to be modified (though much of
it only moved). While doing so, a few minor issues were discovered and
also fixed:

* The check `if len(name) == 0 || len(address) == 0` was used to filter
  out "external" participants, but it also filtered out "email"
  participants that happened to not use a name in their From-header;
  fixed by checking the participant type as stored in the DB instead
* In the query for the event participants, the name of "email"
  participants was `Scan()`ned into a `string`, but it can be `NULL`;
  fixed by coalescing with empty string
* Removed the last usage of Squirrel in this file