api/trackers/export: Add missing nil checks
There are currently two cases that cause the export function to panic.

The first one happens, when the authenticated user does not have the
permissions to access the tracker. In that case the TrackersByID loader
returns nil, so we have to check for that.

The second one can happen for tickets that got mentioned by other
tickets. In that case event.Participant is nil. That is caused here in
tickets.py [1] where participant is only filled for user and
not for ticket mention events.

[1]: https://git.sr.ht/~sircmpwn/todo.sr.ht/tree/c2a87f590d044acaa916cacf42684249f9521f50/item/todosrht/tickets.py#L283